+++ HISTORY.rst
@@ -1,8 +1,18 @@
 History
 =======
 
-1.0dev (unreleased)
--------------------
+1.1.dev0
+--------
+
+- Provide all available constructor options in blueprint.
+  [rnix, 2014-06-19]
+
+- Update select widget to 3.5.0.
+  [rnix, 2014-06-19]
+
+
+1.0
+---
 
 - Make it work.
   [thet]